add MCTS snake suspicious salami
spaceheater will also fall back to using MCTS in case
it detects a situation where the paranoid minimax will
lead to certain death.

+++ src/logic/scoring/floodfill.rs
@@ -4,7 +4,7 @@ use super::{kills, turns_survived};
 
 #[derive(Copy, Ord, Clone, PartialEq, Eq, Default)]
 pub struct VoronoiScore {
-    turns_survived: usize,
+    turns_survived: i64,
     tiles_controlled: usize,
     kills: usize,
     length: isize,
